Partial controllers code and tests.
NOTE: unit tests are incomplete and not working properly. Discovered
that the cascade delete on models is set up wrong, I think deleting one
bridge record is taking much of the database, and lots of other tests
fail due to that. Commented the test that's causing the problem, will
try to fix it tomorrow. zzzzzz